requirements.txt
certifi==2024.7.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.15
django-rest-swagger==2.1.2
djangorestframework==3.15.2
idna==3.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==2.0
openapi-codec==1.3.2
pytz==2018.4
requests==2.31.0
simplejson==3.13.2
uritemplate==3.0.0
urllib3==2.0.7
